+++ backend/internal/transport/http/bloodsearch_handler.go
@@ -20,20 +20,21 @@ import (
 
 // BloodRequestHandler обрабатывает HTTP запросы для операций с заявками на поиск крови
 type BloodRequestHandler struct {
-	createHandler          *bloodcmd.CreateRequestHandler
-	updateHandler          *bloodcmd.UpdateRequestHandler
-	deleteHandler          *bloodcmd.DeleteRequestHandler
-	getByIDHandler         *bloodquery.GetByIDHandler
-	getByPetIDHandler      *bloodquery.GetByPetIDHandler
-	getDonorByIDHandler    *bloodquery.GetDonorByIDHandler
-	getDonationHandler     *bloodquery.GetDonationHandler
-	applyResponseHandler   *bloodcmd.ApplyResponseHandler
-	confirmDonationHandler *bloodcmd.ConfirmDonationHandler
-	rejectDonationHandler  *bloodcmd.RejectDonationHandler
-	closeRequestHandler    *bloodcmd.CloseRequestHandler
-	bloodRequestMapper     *mapper.BloodRequestMapper
-	petMapper              *mapper.PetMapper
-	storage                filestorage.Repository
+	createHandler              *bloodcmd.CreateRequestHandler
+	updateHandler              *bloodcmd.UpdateRequestHandler
+	deleteHandler              *bloodcmd.DeleteRequestHandler
+	getByIDHandler             *bloodquery.GetByIDHandler
+	getByPetIDHandler          *bloodquery.GetByPetIDHandler
+	getDonorByIDHandler        *bloodquery.GetDonorByIDHandler
+	getDonationHandler         *bloodquery.GetDonationHandler
+	applyResponseHandler       *bloodcmd.ApplyResponseHandler
+	confirmDonationHandler     *bloodcmd.ConfirmDonationHandler
+	rejectDonationHandler      *bloodcmd.RejectDonationHandler
+	closeRequestHandler        *bloodcmd.CloseRequestHandler
+	notificationRespondHandler *bloodcmd.NotificationRespondHandler
+	bloodRequestMapper         *mapper.BloodRequestMapper
+	petMapper                  *mapper.PetMapper
+	storage                    filestorage.Repository
 }
 
 func NewBloodRequestHandler(
@@ -48,23 +49,25 @@ func NewBloodRequestHandler(
 	confirmDonationHandler *bloodcmd.ConfirmDonationHandler,
 	rejectDonationHandler *bloodcmd.RejectDonationHandler,
 	closeRequestHandler *bloodcmd.CloseRequestHandler,
+	notificationRespondHandler *bloodcmd.NotificationRespondHandler,
 	storage filestorage.Repository,
 ) *BloodRequestHandler {
 	return &BloodRequestHandler{
-		createHandler:          createHandler,
-		updateHandler:          updateHandler,
-		deleteHandler:          deleteHandler,
-		getByIDHandler:         getByIDHandler,
-		getByPetIDHandler:      getByPetIDHandler,
-		getDonorByIDHandler:    getDonorByIDHandler,
-		getDonationHandler:     getDonationHandler,
-		applyResponseHandler:   applyResponseHandler,
-		confirmDonationHandler: confirmDonationHandler,
-		rejectDonationHandler:  rejectDonationHandler,
-		closeRequestHandler:    closeRequestHandler,
-		bloodRequestMapper:     mapper.NewBloodRequestMapper(storage),
-		petMapper:              mapper.NewPetMapper(storage),
-		storage:                storage,
+		createHandler:              createHandler,
+		updateHandler:              updateHandler,
+		deleteHandler:              deleteHandler,
+		getByIDHandler:             getByIDHandler,
+		getByPetIDHandler:          getByPetIDHandler,
+		getDonorByIDHandler:        getDonorByIDHandler,
+		getDonationHandler:         getDonationHandler,
+		applyResponseHandler:       applyResponseHandler,
+		confirmDonationHandler:     confirmDonationHandler,
+		rejectDonationHandler:      rejectDonationHandler,
+		closeRequestHandler:        closeRequestHandler,
+		notificationRespondHandler: notificationRespondHandler,
+		bloodRequestMapper:         mapper.NewBloodRequestMapper(storage),
+		petMapper:                  mapper.NewPetMapper(storage),
+		storage:                    storage,
 	}
 }
 
@@ -169,6 +172,15 @@ func (h *BloodRequestHandler) Register(api huma.API) {
 		Description: "Закрывает заявку на поиск крови по ID заявки",
 		Tags:        []string{"blood-request-v1"},
 	}, h.CloseBloodSearch)
+
+	huma.Register(api, huma.Operation{
+		OperationID: "respond-to-notification",
+		Method:      http.MethodPost,
+		Path:        "/v1/blood-request/notification/respond/{req_id}",
+		Summary:     "Ответить на уведомление",
+		Description: "Отвечает на уведомление",
+		Tags:        []string{"blood-request-v1"},
+	}, h.NotificationRespond)
 
 }
 
@@ -372,4 +384,16 @@ func (h *BloodRequestHandler) CloseBloodSearch(ctx context.Context, input *commo
 		return nil, err
 	}
 	return &commondto.DefaultMessageOutput{Body: commondto.ResultMessage{Message: "Заявка успешно закрыта, все невыполненные донации отменены"}}, nil
+}
+
+func (h *BloodRequestHandler) NotificationRespond(ctx context.Context, input *dto.NotificationRespondInput) (*commondto.DefaultMessageOutput, error) {
+	if err := h.notificationRespondHandler.Handle(ctx, bloodcmd.NotificationRespondInput{
+		BloodRequestID: input.ID,
+		Action:         bloodcmd.NotificationAction(input.Body.Action),
+	}); err != nil {
+		return nil, err
+	}
+	return &commondto.DefaultMessageOutput{Body: commondto.ResultMessage{
+		Message: "Ответ принят",
+	}}, nil
 }